# docs/openapi/serve.yaml
openapi: 3.0.3
info:
  title: sharecli serve
  description: |
    Minimal OpenAPI stub for the `sharecli serve` HTTP surface.
    Paths mirror `src/commands/serve.rs` (liveness, readiness, Prometheus, config).
  version: 0.3.0
servers:
  - url: http://127.0.0.1:9000
    description: Default local bind (`sharecli serve --bind 127.0.0.1:9000`)
paths:
...
  /debug/pprof/profile:
    get:
      operationId: pprofProfile
      summary: CPU flamegraph sample (opt-in)
      description: |
        Continuous-profiling surface (audit-v38 L45). Requires `SHARECLI_PPROF=1`.
        Unix builds return an SVG flamegraph; Windows returns 501 (use samply/perf).
        Not a public path — Bearer required when AuthN is enabled.
      security:
        - bearerAuth: []
      parameters:
        - name: seconds
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 60
            default: 10
          description: Capture window in seconds (capped at 60).
        - $ref: "#/components/parameters/Traceparent"
      responses:
        "200":
          description: Flamegraph SVG
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        "404":
          description: Profiling disabled (`SHARECLI_PPROF` unset/`0`)
        "501":
          description: Platform does not embed pprof (Windows)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Optional. When `SHARECLI_SERVE_TOKEN` (or `config.serve.bearer_token`) is
        set, required on all paths except `/healthz` and `/readyz`.
  parameters:
    Traceparent:
      name: traceparent
      in: header
      required: false
      schema:
        type: string
        example: 00-0af7651916cd43dd8448eb211c80319c-b7ad6b7169203331-01
      description: W3C Trace Context; extracted into the `http.request` span and echoed on the response.
...